+++ packages/web-core/src/factories/event.ts
@@ -5,10 +5,15 @@ import {
   ControlEventAction,
   DataChunkDataType,
   InworldPacket as ProtoPacket,
+  MutationEvent,
   Routing,
+  SessionControlEvent,
   TextEventSourceType,
 } from '../../proto/ai/inworld/packets/packets.pb';
-import { CancelResponsesProps } from '../common/data_structures';
+import {
+  CancelResponsesProps,
+  SessionControlProps,
+} from '../common/data_structures';
 import { protoTimestamp } from '../common/helpers';
 import { Character } from '../entities/character.entity';
 import { TriggerParameter } from '../entities/inworld_packet.entity';
@@ -163,6 +168,53 @@ export class EventFactory {
     };
   }
 
+  static sessionControl(props: SessionControlProps): ProtoPacket {
+    const sessionControl = {
+      ...(!!props.capabilities && {
+        capabilitiesConfiguration: props.capabilities,
+      }),
+      ...(!!props.sessionConfiguration && {
+        sessionConfiguration: props.sessionConfiguration,
+      }),
+      ...(!!props.clientConfiguration && {
+        clientConfiguration: props.clientConfiguration,
+      }),
+      ...(!!props.userConfiguration && {
+        userConfiguration: props.userConfiguration,
+      }),
+      ...(!!props.continuation && { continuation: props.continuation }),
+      ...(!!props.sessionHistory && {
+        sessionHistoryRequest: props.sessionHistory,
+      }),
+    } as SessionControlEvent;
+
+    return {
+      packetId: {
+        packetId: v4(),
+      },
+      timestamp: protoTimestamp(),
+      routing: this.openSessionRouting(),
+      sessionControl,
+    };
+  }
+
+  static loadScene(name: string): ProtoPacket {
+    const mutation = {
+      loadScene: {
+        name,
+      },
+    } as MutationEvent;
+
+    return {
+      packetId: {
+        packetId: v4(),
+      },
+      timestamp: protoTimestamp(),
+      routing: this.openSessionRouting(),
+      mutation,
+    };
+  }
+
   baseProtoPacket({
     utteranceId = true,
     interactionId = true,
@@ -203,5 +255,12 @@ export class EventFactory {
         })),
       };
     }
+  }
+
+  private static openSessionRouting(): Routing {
+    return {
+      source: { type: ActorType.PLAYER },
+      target: { type: ActorType.WORLD },
+    };
   }
 }