+++ packages/web-core/src/services/connection.service.ts
@@ -1,11 +1,4 @@
-import {
-  LoadSceneResponseAgent,
-  PreviousState,
-} from '../../proto/ai/inworld/engine/world-engine.pb';
-import {
-  ClientRequest,
-  LoadSceneResponse,
-} from '../../proto/ai/inworld/engine/world-engine.pb';
+import { ClientRequest } from '../../proto/ai/inworld/engine/world-engine.pb';
 import { InworldPacket as ProtoPacket } from '../../proto/ai/inworld/packets/packets.pb';
 import {
   AudioSessionState,
@@ -36,10 +29,10 @@ import {
 import { Character } from '../entities/character.entity';
 import { SessionContinuation } from '../entities/continuation/session_continuation.entity';
 import { InworldPacket } from '../entities/inworld_packet.entity';
+import { Scene } from '../entities/scene.entity';
 import { SessionToken } from '../entities/session_token.entity';
 import { EventFactory } from '../factories/event';
 import { StateSerializationService } from './pb/state_serialization.service';
-import { WorldEngineService } from './pb/world_engine.service';
 
 interface ConnectionProps<InworldPacketT, HistoryItemT> {
   name?: string;
@@ -72,20 +65,18 @@ export class ConnectionService<
   private audioSessionParams: SendPacketParams = {};
   private ttsPlaybackAction = TtsPlaybackAction.UNKNOWN;
 
-  private scene: LoadSceneResponse;
+  private scene: Scene | undefined;
   private session: SessionToken;
-  private connection: Connection<InworldPacketT>;
+  private connection: Connection<InworldPacketT, HistoryItemT>;
   private connectionProps: ConnectionProps<InworldPacketT, HistoryItemT>;
 
-  private characters: Array<Character> = [];
-
+  private eventFactory = new EventFactory();
   private intervals: NodeJS.Timeout[] = [];
+  private packetQueue: QueueItem<InworldPacketT>[] = [];
+
   private disconnectTimeoutId: NodeJS.Timeout;
 
-  private eventFactory = new EventFactory();
-
   private stateService = new StateSerializationService();
-  private engineService = new WorldEngineService<InworldPacketT>();
 
   private onDisconnect: (() => Awaitable<void>) | undefined;
   private onError: (err: Event | Error) => Awaitable<void>;
@@ -173,27 +164,51 @@ export class ConnectionService<
     return this.history.getTranscript();
   }
 
-  async loadCharacters() {
-    await this.loadScene();
+  async getCharacters() {
+    await this.open();
+
+    return this.scene?.characters ?? [];
   }
 
   async open() {
+    if (this.state !== ConnectionState.INACTIVE) return;
+
     try {
-      await this.loadScene();
+      await this.loadToken();
 
-      if (this.state === ConnectionState.LOADED) {
-        this.state = ConnectionState.ACTIVATING;
+      const { client, name, sessionContinuation, user } = this.connectionProps;
 
-        const packets = this.getPacketsToSentOnOpen();
+      const packets = this.getPacketsToSentOnOpen();
 
-        await this.connection.open({
-          session: this.session,
-          convertPacketFromProto: this.extension.convertPacketFromProto,
-          ...(packets.length && { packets }),
-        });
+      this.packetQueue = [...packets, ...this.packetQueue];
+      this.scene = await this.connection.openSession({
+        client,
+        name,
+        sessionContinuation,
+        user,
+        session: this.session,
+        convertPacketFromProto: this.extension.convertPacketFromProto,
+      });
 
-        this.scheduleDisconnect();
+      const factory = this.getEventFactory();
+
+      if (
+        !this.connectionProps.config.capabilities.multiAgent &&
+        !factory.getCurrentCharacter() &&
+        this.scene.characters[0]
+      ) {
+        factory.setCurrentCharacter(this.scene.characters[0]);
       }
+
+      factory.setCharacters(this.scene.characters);
+
+      if (this.scene?.history?.length) {
+        this.setPreviousState(this.scene.history);
+      }
+
+      await this.onReady?.();
+      this.releaseQueue();
+      this.scheduleDisconnect();
     } catch (err) {
       this.onError(err);
     }
@@ -246,36 +261,6 @@ export class ConnectionService<
     }
   }
 
-  private setCharacterList() {
-    const characters = (this.scene?.agents || [])?.map(
-      (agent: LoadSceneResponseAgent) =>
-        new Character({
-          id: agent.agentId,
-          resourceName: agent.brainName,
-          displayName: agent.givenName,
-          assets: {
-            avatarImg: agent.characterAssets.avatarImg,
-            avatarImgOriginal: agent.characterAssets.avatarImgOriginal,
-            rpmModelUri: agent.characterAssets.rpmModelUri,
-            rpmImageUriPortrait: agent.characterAssets.rpmImageUriPortrait,
-            rpmImageUriPosture: agent.characterAssets.rpmImageUriPosture,
-          },
-        }),
-    );
-
-    const factory = this.getEventFactory();
-
-    if (
-      !this.connectionProps.config.capabilities.multiAgent &&
-      !factory.getCurrentCharacter() &&
-      characters[0]
-    ) {
-      factory.setCurrentCharacter(characters[0]);
-    }
-
-    factory.setCharacters(characters);
-  }
-
   private async write(getPacket: () => ProtoPacket) {
     let inworldPacket: InworldPacketT;
 
@@ -295,10 +280,7 @@ export class ConnectionService<
         this.intervals.push(interval);
       });
 
-    // 1. Send a packet to a connection.
-    // The packet will be sent directly or added to the queue.
-    // If the connection is not active, we need to add the packet to the queue first to guarantee the order of packets.
-    this.connection.write({
+    const itemToSend = {
       getPacket,
       afterWriting: (packet: InworldPacketT) => {
         inworldPacket = packet;
@@ -312,54 +294,27 @@ export class ConnectionService<
           await this.interruptByPacket(packet);
         }
       },
-    });
+    };
 
-    // 2. Open the connection if it's inactive.
-    if (!this.isActive()) {
-      this.open();
+    if (this.isActive()) {
+      this.connection.write(itemToSend);
+    } else {
+      this.packetQueue.push(itemToSend);
+
+      await this.open();
     }
 
     return resolvePacket();
   }
 
-  private async loadScene() {
-    if (this.state === ConnectionState.LOADING) return;
+  private async loadToken() {
+    if (this.state === ConnectionState.ACTIVATING) return;
 
-    const { name, client, user } = this.connectionProps;
-
-    try {
-      await this.ensureSessionToken({
-        beforeLoading: () => {
-          this.state = ConnectionState.LOADING;
-        },
-      });
-
-      if (!this.scene) {
-        this.scene = await this.engineService.loadScene({
-          config: this.connectionProps.config,
-          session: this.session,
-          sessionContinuation: this.connectionProps.sessionContinuation,
-          extension: this.extension,
-          name,
-          user,
-          client,
-        });
-
-        if (this.connectionProps?.config?.history?.previousState) {
-          this.setPreviousState(this.scene?.previousState);
-        }
-
-        this.setCharacterList();
-      }
-
-      if (
-        [ConnectionState.LOADING, ConnectionState.INACTIVE].includes(this.state)
-      ) {
-        this.state = ConnectionState.LOADED;
-      }
-    } catch (err) {
-      this.onError(err);
-    }
+    await this.ensureSessionToken({
+      beforeLoading: () => {
+        this.state = ConnectionState.ACTIVATING;
+      },
+    });
   }
 
   async ensureSessionToken(props?: { beforeLoading: () => void }) {
@@ -394,17 +349,14 @@ export class ConnectionService<
     }
   }
 
-  private setPreviousState(previousState: PreviousState) {
-    const { stateHolders = [] } = previousState || {};
-    stateHolders.forEach((stateHolder) => {
-      stateHolder.packets?.forEach((packet) =>
-        this.history.addOrUpdate({
-          grpcAudioPlayer: this.connectionProps.grpcAudioPlayer,
-          characters: this.eventFactory.getCharacters(),
-          packet: this.extension.convertPacketFromProto(packet),
-        }),
-      );
-    });
+  private setPreviousState(previousPackets: ProtoPacket[]) {
+    previousPackets.forEach((packet) =>
+      this.history.addOrUpdate({
+        grpcAudioPlayer: this.connectionProps.grpcAudioPlayer,
+        characters: this.eventFactory.getCharacters(),
+        packet: this.extension.convertPacketFromProto(packet),
+      }),
+    );
 
     const changed = this.history.get();
 
@@ -417,6 +369,23 @@ export class ConnectionService<
     if (this.disconnectTimeoutId) {
       clearTimeout(this.disconnectTimeoutId);
     }
+  }
+
+  private releaseQueue() {
+    this.packetQueue.forEach((item: QueueItem<InworldPacketT>) =>
+      this.connection.write(item),
+    );
+
+    this.packetQueue = [];
+  }
+
+  private clearQueue() {
+    this.intervals.forEach((i: NodeJS.Timeout) => {
+      clearInterval(i);
+    });
+
+    this.intervals = [];
+    this.packetQueue = [];
   }
 
   private initializeHandlers() {
@@ -437,7 +406,7 @@ export class ConnectionService<
         webRtcLoopbackBiDiSession.getPlaybackLoopbackStream(),
       );
       this.state = ConnectionState.ACTIVE;
-      onReady?.();
+      await onReady?.();
     };
 
     this.onDisconnect = async () => {
@@ -618,14 +587,6 @@ export class ConnectionService<
     if (changed.length) {
       this.connectionProps.onHistoryChange?.(this.getHistory(), changed);
     }
-  }
-
-  private clearQueue() {
-    this.intervals.forEach((i: NodeJS.Timeout) => {
-      clearInterval(i);
-    });
-
-    this.intervals = [];
   }
 
   private getPacketsToSentOnOpen() {